To revert: failed attempt to shrink TreeNodes by removing the tx channel
Unfortunately this doesn't work because you deadlock since there are Senders always kept around. You could potentially use `try_recv` in the worker threads, but then you need some root type that destroys the root worker thread's sender and that requires another channel to send a message to the root worker thread (since we need to be executing in its thread), but that sounds very complicated and unpleasant. I'll fix this once lockness is out.

Signed-off-by: Alex Saveau <saveau.alexandre@gmail.com>