Avoid recursion in drop impl most of the time
As a very nice bonus, this fixes the explosion of erroring syscalls. Previously, if an error occurred in a leaf node, it would bubble all the way up through the tree, attempting to delete the parent directories along the way which clearly won't work since they're not empty. Now we'll only get two failures: the original one and then the one from the adjacent directory trying to finish its deletion.

Signed-off-by: Alex Saveau <saveau.alexandre@gmail.com>