Undo 7ce73a4c610b3f4dec0ff036fcbaed5650d9bcdc as it broke recursive copy detection for multiple directories
Example:
$ ^mkdir -p a/b/c
$ cpz a a/b a/b/c/

If you copy only one of the directories it works, but not both since we pick a global inode for all copies instead of one per root task.

Note that the example above is still totally broken because you're modifying the directory in parallel with the copy of itself. Separately, you can still break things by creating a symlink that points to its parent. Solving both these problems requires building the file tree in-memory which isn't worth it.

Signed-off-by: Alex Saveau <saveau.alexandre@gmail.com>